Started Player

+++ src/main/java/fr/unicorn/lumiobase/models/Scenario/ScenarioRepository.java
@@ -2,5 +2,5 @@ package fr.unicorn.lumiobase.models.Scenario;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScenarioRepository extends CrudRepository<Scenario, Integer> {
+public interface ScenarioRepository extends CrudRepository<Scenario, Long> {
 }
